+++ docs/13_0_0/gettingstarted/whatsnew.md
@@ -30,6 +30,8 @@ This page contains a list of big features. Please check the GitHub issues for al
     * Added attribute `exportTag` and `exportRowTag` to allow finer grain control over XML export.
 * Schedule
     * Added event `eventDblSelect` to allow an event to be double clicked.
+* SelectCheckboxMenu
+    * Added attribute `var` to support custom content in overlay panel.
 * SpeedDial
     * Added attribute `ariaLabel` to allow screen reader support on button and `title` for tooltip.
 * Menu